+++ sdk/api/conn-sessions-mgr-api.go
@@ -38,7 +38,7 @@ type ISessionsMgrApi interface {
 	CreateSession(
 		tx *sql.Tx,
 		ctx context.Context,
-		devId int32,
+		devId int64,
 		sessionType string,
 		timeSecs int,
 		dataMbytes float64,
@@ -46,7 +46,7 @@ type ISessionsMgrApi interface {
 		downMbits int,
 		upMbits int,
 		useGlobal bool,
-	) (int32, error)
+	) (int64, error)
 
 	// Get the current running session of a client device.
 	CurrSession(clnt IClientDevice) (cs IClientSession, ok bool)
